+++ libs/model/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_library(model
     model.cpp
+    model_crypto_provider_impl.cpp
+    model_hash_provider_impl.cpp
     model_crypto_provider_stub.cpp
     )
